==> scripts/00-install_packages.R <==
#### Preamble ####
# Purpose: Install necessary packages for the analysis
# Author: Denise Chang
# Date: 1 November 2024
# Contact: dede.chang@mail.utoronto.ca
# License: MIT
# Pre-requisites: None

#### Install Packages ####

# install necessary packages to run the files in this repository.
install.packages("tidyverse")
install.packages("arrow")
install.packages("here")
install.packages("janitor")
install.packages("DiagrammeR")
install.packages("DiagrammeRsvg")
install.packages("rsvg")
install.packages("ggfortify")
install.packages("psych")
install.packages("kableExtra")
install.packages("plotly")
install.packages("RColorBrewer")
install.packages("modelsummary")
install.packages("broom")
install.packages("sf")
install.packages("patchwork")
install.packages("testthat")
install.packages("readr")
